Code updated

+++ S04T02N03/src/main/java/cat/itacademy/s04/t02/n03/services/FruitService.java
@@ -1,5 +1,7 @@
 package cat.itacademy.s04.t02.n03.services;
 
+import cat.itacademy.s04.t02.n03.exceptions.FruitDoesNotExistException;
+import cat.itacademy.s04.t02.n03.exceptions.FruitNotFoundException;
 import cat.itacademy.s04.t02.n03.model.Fruit;
 import cat.itacademy.s04.t02.n03.repository.FruitRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,15 +24,22 @@ public class FruitService {
     }
 
     public Fruit updateFruit(Fruit fruit) {
+        if (!repository.existsById(fruit.getId())) {
+            throw new FruitDoesNotExistException("Fruit with ID " + fruit.getId() + " does not exist.");
+        }
         return repository.save(fruit);
     }
 
     public void deleteFruit(String id) {
+        if (!repository.existsById(id)) {
+            throw new FruitDoesNotExistException("Cannot be removed: fruit with ID " + id + " does not exist.");
+        }
         repository.deleteById(id);
     }
 
-    public Optional<Fruit> getFruitById(String id) {
-        return repository.findById(id);
+    public Fruit getFruitById(String id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new FruitNotFoundException("Fruit with ID " + id + " not found."));
     }
 
     public List<Fruit> getAllFruit() {
